requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.23.2
pydantic==2.6.2
python-dotenv==1.0.0
requests==2.31.0
openai==1.32.0
